Add a background texture
This adds a new background texture, thanks to http://subtlepatterns.com/
which contains various awesome background patterns.

Overlayed the colour that I had picked out over top of it and made the
opacity just subtle enough so that it is not distracting, but is still
there breaking up the dullness...

I like it.